Extract the navigation link to a useNavigationLink hook
